Allow burning without foundry

+++ src/account/operations/helpers/time.rs
@@ -26,7 +26,7 @@ impl AccountHandle {
             .expect("Time went backwards")
             .as_secs();
         let status_response = self.client.get_info().await?.nodeinfo.status;
-        let latest_ms_timestamp = status_response.latest_milestone_timestamp;
+        let latest_ms_timestamp = status_response.latest_milestone_timestamp as u64;
         // Check the local time is in the range of +-5 minutes of the node to prevent locking funds by accident
         if !(latest_ms_timestamp - FIVE_MINUTES_IN_SECONDS..latest_ms_timestamp + FIVE_MINUTES_IN_SECONDS)
             .contains(&local_time)
